test(close-button): recategorize and clean up browser tests
